Fix getInterval return type

+++ jmri/CommonTurnoutOperation.java
@@ -7,7 +7,7 @@ package jmri;
  * CommonTurnoutOperation class - specialization of TurnoutOperation to contain
  * common properties and methods for concrete subclasses
  * @author John Harper
- * @version $Revision: 1.3 $
+ * @version $Revision: 1.4 $
  */
 public abstract class CommonTurnoutOperation extends TurnoutOperation {
 
@@ -41,7 +41,7 @@ public abstract class CommonTurnoutOperation extends TurnoutOperation {
 	 */
 	public abstract TurnoutOperator getOperator(AbstractTurnout t);
 	
-	public long getInterval() { return interval; };
+	public int getInterval() { return interval; };
 	
 	public int getMaxTries() { return maxTries; };
